+++ GithubApiDemo/app/src/main/AndroidManifest.xml
@@ -5,7 +5,7 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name=".App"
+        android:name=".base.App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -13,10 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.GithubApiDemo">
         <activity
-            android:name=".UserDetailActivity"
-            android:exported="true" />
-        <activity
-            android:name=".UserListActivity"
+            android:name=".function.UserListActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -24,6 +21,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".function.UserDetailActivity"
+            android:exported="true" />
     </application>
 
 </manifest>
