+++ GithubApiDemo/app/src/main/java/com/caizhixng/githubapidemo/base/App.kt
@@ -1,4 +1,4 @@
-package com.caizhixng.githubapidemo
+package com.caizhixng.githubapidemo.base
 
 import android.app.Application
