+++ GithubApiDemo/app/src/main/java/com/caizhixng/githubapidemo/net/Page.kt
@@ -1,9 +1,10 @@
-package com.caizhixng.githubapidemo
+package com.caizhixng.githubapidemo.net
 
 /**
  * czx 2021/9/11
  */
 data class Page(var page: Int, val perPage: Int, var keyWord: String) {
+
     private val startPage = 1
 
     fun addPage() {
