+++ GithubApiDemo/app/src/main/res/layout/activity_user_detail.xml
@@ -4,7 +4,7 @@
     android:id="@+id/swipe"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UserDetailActivity">
+    tools:context=".function.UserDetailActivity">
 
     <WebView
         android:id="@+id/webview"
